refactor: 커스텀 에러 바깥으로 추출
